Conflicts:
	Ascent/Assets/Scripts/Character/Enemy/Enemy.cs

+++ Ascent/Assets/Scripts/AI/AICondition_Attacked.cs
@@ -18,7 +18,7 @@ public class AICondition_Attacked : AICondition
         return conditionMet;
     }
 
-    public void OnHit(float f)
+    public void OnHit(int f)
     {
         wasHit = true;
     }
